pyproject.toml
[project]
name = "street-art-bot"
version = "0.1.0"
description = "Telegram bot for street art walking tours (SpotBot / Место)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot==21.9",
    "motor==3.6.0",
    "pydantic==2.10.6",
    "pydantic-settings==2.7.1",
]

[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bot"]
